Modify get_md5_tasks to skip tasks with no vmcores or invalid md5sum files
It is possible tasks fail due to various reasons or the md5sum is
not fully calculated (in which case we may have 'Enabled' as the
contents of md5sum file).  In these cases it is best to avoid adding
to the list of tasks with md5sum as if we take these tasks, we end
up rejecting later and creating lines in the retrace_log files of
tasks such as:
2018-06-21 10:00:03 Attempt to dedup /cores/retrace/tasks/870526745/crash/vmcore and /cores/retrace/tasks/496319104/crash/vmcore but 'stat' failed on one of the paths

Check to see if the vmcore exists and check the md5sum file
contents for a valid md5sum value.  If any of the new checks
fail, skip over the task.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>